test(github-client): error-handler契約テストを別ファイルへ分離
【背景】error-handler の直接契約テスト3件が facade テストファイルに残っており、
検証対象モジュールとテストファイルの対応が崩れていた。

【変更内容】GitHubAPIError の基底クラス契約と SanitizedJSONDecodeError の
__str__ / __reduce__ 契約を検証する3件を、所有モジュールに対応する
test_github_error_handler.py へ移動。APIClientError の import も随伴移動。
AST 実測で移動3件・残留14件・移動先既存54件すべて byte-identical、
関数名集合71件の保存を確認した純粋移動。

品質ゲート: 1443 tests passed / 97.60% coverage / ruff 0 errors / mypy 0 errors
